LayoutTests/fast/css/pseudo-default-004.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../js/resources/js-test-pre.js"></script>
<style>
 input { background: lime; }
 button { background: lime; }
 :default { background: red; }
</style>
</head>
<body>
<p id="description"></p>
<form method="get">
<input name="victim" type="button" value="Not a submit" />
<input name="victim" type="button" value="Not a submit"/>
<button name="victim" type="button"/>Not a submit</button>
</form>
<div id="console"></div>
<script>
description("This test performs a simple check for the :default CSS selector: it shouldn't be applied.");

v = document.getElementsByName("victim");
for (i = 0; i < v.length; i++)
    shouldBe("document.defaultView.getComputedStyle(v[i], null).getPropertyValue('background-color')", "'rgb(0, 255, 0)'");
</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
