third_party/WebKit/LayoutTests/fast/css/pseudo-default-004-expected.txt

This test performs a simple check for the :default CSS selector: it shouldn't be applied.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

  Not a submit
PASS document.defaultView.getComputedStyle(v[i], null).getPropertyValue('background-color') is 'rgb(0, 255, 0)'
PASS document.defaultView.getComputedStyle(v[i], null).getPropertyValue('background-color') is 'rgb(0, 255, 0)'
PASS document.defaultView.getComputedStyle(v[i], null).getPropertyValue('background-color') is 'rgb(0, 255, 0)'
PASS successfullyParsed is true

TEST COMPLETE
